Sysop and user tests
Some tests require a user or sysop, but are not yet flagged
as user=True.

This changeset also adds a sysop flag, to identify tests which
require a sysop account, skipping the test if this condition is
not met.

All tests except pwb.py-based tests can now be run without
user-config.py, with tests skipping if they require a username.

Change-Id: I690423556f1266f3ebe0bee11aa91b41b9965baf