[FIX] Only apply uppercase to Link title if namespace dictates so
The namespace defines if a Link title begins with an uppercase character.
The main namespace of Wiktionary, for example requires an exact match. As
the namespace always define the case, using the site case isn't necessary.

If the namespace doesn't report the case property it choses the site global
case property.

Bug: 69118 (partially)
Change-Id: If86346dc203c54d2e6631e3dddd106add3c34d30