Fix newitem.py to instantiate a new item
Ibea8f84242a6e938882059f5d5fa28394b8a7a23 changed the ItemPage.fromPage
semantics to raise NoPage if the page does not exist, or is not linked.

Catch this exception, and instantiate a new item to call editEntity with.
Also add test case to create a new item on testwikidata.

Bug: 71095
Change-Id: I612f56d844684d85890af1c7d33051c292552253