[IMPROV] Use str.decode instead of unicode
With Python 3 unicode was removed. To support both versions 'unicode' is
aliased as 'str' in Python 3. But the usage of 'unicode(str, enc)' is
not necessary and instead 'str.decode(enc)' can be used (or to be exact
it's 'bytes' instead of 'str' in Python 3). Unfortunately this does not
remove the need for the 'unicode' alias, but reduces it a bit.

Change-Id: I7118142b7881821d78e4cb848116c4a2e5721c0a