[FEAT] Deprecrate module classes and variables
This adds a way to deprecrate classes or variables in modules. The
difference to '@deprecated' is that the access alone is important and
not calling the function. So this warns also if someone is just
importing a class (e.g. 'from pywikibot.excptions import
UploadWarning').

Unfortunately __getattr__ is not available on module level, so the
module must be wrapped in a class which redirects all accesses.

Change-Id: I3e83472ba780d9d99745c8dff864926432e9f254