Page: ValueError with site as source and empty title
Created BasePage() class in order to allow different behaviour in
subclasses Page and WikibasePage.

Page() will raise ValueError if no or empty string is provided as
title when Site is used as source, as api.php cannot handle an
empty string as title.

WikibasePage() will be able to be instantiated with minimal
constructor values.

Also fixed pep257 errors in:
- tests/page_tests.py
- tests/wikibase_tests.py

Bug 60381

Change-Id: Iba0d3d7a53a434301fbea4261a0efda24ffd8af1