Use Site interface in cache key
pywikibot.Site is invoked with different values for 'interface',
notably 'DataSite' is used by APISite.data_repository().

If a DataSite is instantiated before an APISite for a wiki, it is
cached, and returned when an APISite object is requested.  DataSite
overrides APISite/BaseSite methods, including raising exceptions for
deprecated methods, so the cached & returned DataSite is not an acceptable
returned value if an APISite was requested.

Bug: 71451
Change-Id: I93195dcba3185c6df233bac89ed981ddccc78abd