Use unicode for Request params
Request.http_params() optionally decodes values in the Request.params dict,
and then always re-encodes params each time it is called.  As a result
the type of the values differs throughout the life of the Request
object.

Split Request.http_params() into __setitem__, _add_defaults, _http_param_string.
Rename Request.params to _params as Request has a dict interface for it.

__setitem__ forces all param values to always be unicode
(or str for ascii in Python 2), and be stored as a list
for each key.

_add_defaults adds param values to the pending request.  It now only
peforms these additions the first invocation, and is a noop when invoked
again.

_http_param_string prepares a URL for the site's API, but does not store
the encoded values in the Request object's state.

Request.action is added as this parameter is mandatory, may only have
one value, and is frequently used in the logic.

Change-Id: Id13622e72623024166cec4ac0a5c9c4b3d511755