@deprecated support for any object
Decorator @deprecated currently assumes it is run on a bound method.
If the decorated function doesnt have an argument 'self' or 'cls',
fallback to reporting it as a function.

Also always report which module the deprecated method is in, as the
user may not recognise BaseSite if they are using many packages, and
pywikibot might be used indirectly by another package.

Change-Id: I24af9784e95e0f57a39dd6523267187ef6a4b8c6