Allow @unittest.expectedFailure on all TestCase methods
TestCase allows test methods to be 'multi-site' tests, which
will be re-run for each site in the TestCase sites matrix.

The metaclass has previously checked that the test method only
has 1 or 2 arguments.  @unittest.expectedFailure replaces the
test method with a decorated version that appears as if it has
zero arguments, due to the way decorators work.  This change
allows test methods with 0 arguments.

Change-Id: Ib6d74af168e0bad93df65da2965daffa6b3951e0