Move typical page save exception handling into Bot
Most bots wish to continue processing pages from their generator
even if a page fails to save, whether because of a page-specific
or server related error.

This functionality has been added to Bot.userPut, not enabled by default.
reflinks.py has been migrated to use this functionality.

Also, add show_diff to Bot.userPut, which can be used to disable
showing the changes that will be made to the page.

Change-Id: I34898b36f915a1399522040201aac84fa8f1324c