Increase limits in QueryGenerator when data are sparse
In QueryGenerator's query loop, instead of just narrowing down limit
given to each api request.submit(), increase limit when only
"query-continue" data are fetched by a request.submit().

This will reduce the number of needed queries if data are "sparse".

See bug details for an example with data showing the background of the change.

Bug: 72209
Change-Id: I800a3399a991f5c51d23ed38b8636f09b704a9c6