[FIX] TokenWallet: Assert a username
The TokenWallet assumed the normal user and not the sysopuser is logged
in. This worked as long as both are named the same, but if a sysopuser
is logged in with a name different from the normal username, it'll fail
and tell that the normal user is not logged in.

It now just checks whether a user is logged in, which is also what the
next statement requires.

Bug: 72238
Change-Id: Iccce554b5c9e83a42ae87c0bd69e38f880e3f6ac