Add cert test to version script
Each platform may patch the httplib2 variable CA_CERTS, and
httplib2 allows this variable to be dynamically patched by
package 'httplib2.ca_certs_locater'.

To simplify debugging of ssl certificate verification errors,
it is useful to know which file is being loaded, and whether it
includes the certificate needed for Wikimedia projects.

Bug 72009

Change-Id: Ib3d1c5e1b5d683204150b04f3df4176754fbaaa5