[FEAT] Request: Support simplified continuation
With version 1.25wmf5 a warning is emitted that the continue parameter
is going to change (af11268d05d98d27d99cfee9264d649a62c2c816). It is
possible to avoid that warning by explicitly telling the server which
continuation mode to expect.

This adds the ability of QueryGenerator to use the simplified
continuation method as soon as possible.

If the server is using a version of at least 1.25wmf5 it'll
automatically 'rawcontinue' to the request if neither 'continue' nor
'rawcontinue' are set. This will avoid any warning emitted and in the
future (if by default it uses the simplified continuation mode) it'll
use the old continuation mode for which the scripts are probably
written. To avoid an infinite loop, because requesting the server
version requires a request which requires (raw-)continue in the
parameters, siteinfo uses the simplified continuation mode by always
adding "continue=''" to the parameters.

Change-Id: Iddbf4922c60c0f3900b88cb81e4c176855e33fc3