Fix doctest broken by deprecation warning
MW API has deprecated the old request query system, issuing a warning
if neither 'continue' nor 'rawcontinue' are in the request params.

This breaks the doctest for the raw Request object, which prevents
jenkins from approving code changes which dont resolve this issue.

Modified the expected outcome to allow for warnings in the response,
so that the dev system isnt in stalemate while waiting for the right
fix.

Change-Id: I07081de95437724233afe38df663256a3e495280