upload.py: Allow to ignore warnings
Added the option -ignorewarn to ignore specific warnings, or all warnings,
using the same logic as the abortonwarn. Useful for batch uploads of files
to replace existing ones.

The UploadRobot already had a boolean ignoreWarning variable in
constructor, which has been extended to allow an array of warning
codes to ignore.

The message shown when a warning is found has been extended to display the
warning code, so the user can know what warning code use to ignore or
abort it.

Change-Id: Idf9e562f5a4fbae8184718c66cedc90042379e81