Remove unused Family.content_id
content_id is used in compat, for non-API access to mediawiki, which
is functionality that exists in core.
If core ever needs to support pre-1.13 editing, this and many other
parts of compat will need to be extensively mined for implementation
details like this.  We know where to find this.  It doesnt need to
be in core Family.

Change-Id: Ib3111eba7c16782a68646f34b540062486e17e6e