Export sys.path to pwb scripts
setup.py can install packages, which are available within the
unit tests as it adds the packages to sys.path.

script_tests needs to export sys.path into the script environment
so these packages may be imported by the script.

Also move a docstring which was incorrectly placed inside the method
procedure.

Bug: 71864
Change-Id: I6356b4b5d8da7529957befe8157f78457ebb5df1