WikiApiary AutoFamily
WikiApiary is on the meta interwiki map ([[meta:IWM]]), and is a
useful dataset for the development of pywikibot, e.g. providing
lists of sites which use MediaWiki generator versions, which can
be used to test pywikibot against specific versions.

It will also be the only family registered in pywikibot that is
Semantic MediaWiki, so it may be useful when testing bugs relating
to Semantic MediaWiki.

Bug: 72854
Change-Id: Id5e7ee43e0ccf944f2375918fee7520c7884bf86