New WikiStats module
wikistats.wmflabs.org provides aggregate statistics for families of
wikis.  Pywikibot currently fetches a subset of the data using XML
to create a hard-coded list of languages by size for the multi-lang
WMF families and family 'anarchopedias'.

This new module provides raw access to all of the data stored in
wikistats, and allows it to be fetched using the smaller csv format,
however the XML format is also supported and is used for Python 2
without the unicodecsv module installed.
e.g. for wiktionaries, the XML is 300Kb vs the csv 26Kb.

Change-Id: Id0070092d2337c9fc86b01e2103999c6dcea42fa