Use envvar EDITOR if it exists, with tests
EDITOR is usually set on Unix platforms to be a basic editor.

Use this as the default editor, if set, even on Windows.

Set EDITOR in the script tests to be a noop (like /bin/true) so
the script continues.

Scripts 'editarticle' and 'upload' are now able to be tested.

Change-Id: I8a069960d0919c760441538c0ac0d11d1015bf01