[FEAT] Site: Search for interwiki prefix for a site
This returns all interwiki prefixes redirecting to a site from the
original site. It is like the reverse of APISite.interwiki and could be
useful to create link texts to other sites, without relying on the fact
that the language could be used for a interwiki link within the same
family.

This also changes the behaviour that now all sites are cached. So when
requesting the site for an interwiki prefix it previously had only
cached that site to that prefix.

Change-Id: Id2b889ccaf3c3f122429d59f5cf064855b0b4fea