[FIX] pagegen tests: apply case sensitivity
When the pagegenerator tests are done with case sensitive namespaces the
tests failed because the title read from the file were lowercase but the
test expected uppercase. This adds the expected results for both case
sensitivity settings.

Also added an test with an explicit uppercase value that those page
doesn't get lowercased by accident.

Change-Id: Ica76e5218f9301b76b537d9c532f8c6c8fac85f7