ParamInfo class
Manages data from the paraminfo which was previous managed within
QueryGenerator and stored in APISite._modules.

This provides preloading of an initial set of params, so the cache
key is constant, and bulk loading of all query params, again so that
the cache key is constant irrespective of order of queries used.

Change-Id: I84a0769f19abf8740106659426109c62e8438e65