Fix remaining pep8 issue: E265: use of comments
E265 is '#' not followed by a space, and is part of the
normal pep8 rules.  Fixed existing problems and made it
part of the mandatory checkin rules.

This was used for:
- visual dividers in code
- commented out debugging code
- commented out code that was replaced
- commented out code instead of FIXME:/TODO:
- commented out example code

and occasionally a normal comment with the space
missing after the #

Replaced some with logging, raised NotImplementedException in
checkimages.py with FIXME and bug numbers.

Change-Id: I02864ad75f9e11f5cd8821fe13f52b57a93d56c3