[FIX] ParamInfo: Don't modify module set
The module set has been modified which led to the effect that when
someone was fetching all entries and then iterating over the entries of
the same set, that all already cached entries were missing from the set.

The HTTP GET patch (c26b9dcb4242add0145a883c6b6973a45a665ed0) therefore
was thinking it could use a GET request, because the entry which
should've shown it that a POST request is required was deleted.

Change-Id: I467ff7a8fe243df0abd7f322ecaa8460a91e4017