[FIX] Upload: Use fake filename in MIME header
The standard MIME request does only support ASCII characters in the
headers. As the value is apparently not interpreted it's replaced by
a dummy value, so Unicode characters in the file page name doesn't cause
pywikibot to receive an API error.

There is RFC2231 which is supported by both Python versions but not by
PHP on which the MediaWiki server rely (see bug: 73662). This does only
fix the uploads as it is unknown how to handle other MIME requests.

Bug: 73661
Change-Id: I5ea92dcf1115b675b56142494d805e75d389b7d4