[FEAT] Replace: Support multiple replacements
This supports multiple -fix and command line replacements at once. Each
fix set and also each fix in that set could define it's own edit summary
(the third value in the tuple).

If no user defined summary is given it will only use the summaries of
the applied replacements. For the replacements given via the command
line it will create one merged replacement message. The rest are added
sorted alphabetically after it.

Change-Id: Ic14a8236d143050b267bee9308006b4d14d8ca6d