Permissive Timestamp fromX methods
Several methods of site and page classes return a Timestamp where
previously they returned a string which should be passed to the
Timestamp class to create a Timestamp object.
In order to improve compatibility with old versions of pywikibot,
Timestamp should recognise it is being passed a Timestamp, and not
try to parse it.

Change-Id: Ifb850b3e52c0fb1b84f2da433f50cb142f854591