[FIX] Allow deprecation of trailing parameters
With 7aa43ba4a0c8e12bda4a62e32e062672ec8851fa the parameters of several
methods have been deprecated leaving just 'self'. Because only one or
two parameters were used before, the possibility is high that they were
used positionally (without the parameter name) and the @deprecated_args
syntax can't support that (reliably).

This adds a decorator to deprecate parameters which appeared at the end
of a method/function. The names of those parameters must be specified
and if that method/function is called with too many parameters those are
removed and a warning is issued.

Bug: T75489
Change-Id: I9035f3b6be86fe7b34e515e19affe8ef1a9a558e