[FIX] Sametitle: Don't discard invalid namespaces
If the namespace was invalid, sametitle would've still just compared the
text after the 'namespace' so a page title like 'Foo:Bar' would result
in 'namespace main' and 'name is Bar' although it should've been 'name
is Foo:Bar' because 'Foo' is not a valid namespace.

Also modified the tests so this will be more visible in the future.

Change-Id: Idd42d63125fa2470374c0392e26cceac9e49bb30