[FEAT] Date: Add/subtract months from a date
This allows to add or subtract any number of months with a calendar that
has 12 months. It can also calculate the difference between two days in
terms of months.

Because Python's datetime module is based around a calendar with 12
months it works directly on it and assumes that a year has always 12
months.

Bug: T73124
Change-Id: I4b204465b7224f4d86e3f4462847002f3382ac3a