Replace cosmetic change 'validXhtml' with 'HTML' fix
MediaWiki 1.16+ $wgWellFormedXml defaults to converting <br> to <br/>
when rendering, making this cosmetic change unnecessary, and undesirable
to be performing it in every save operation.

Re-adding it as part of the 'HTML' fix which can be applied as desired,
e.g. versions prior to 1.16, cleaning the underlying wikitext.  It was
disabled from the HTML fix in compat dcf27d0c:
    https://www.mediawiki.org/wiki/Special:Code/pywikipedia/991

Bug: 55236
Change-Id: I27789904778c6d86dc6897d88215939f22fe68ca