[IMPROV] Reduce usage of unicode()
The unicode() builtin is not necessary if it's added unmodified via
format or %-notation into a unicode string (e.g. u'{0}'.format(value)).

In addition, the localized time formatting was removed. The default 'C'
locale (which is ASCII by definition) is now used instead.

If the parameter is a list or tuple it's using format to avoid problems
with Python that it interprets the tuple as multiple values. Lists too,
as they might be replaced by tuples (who knows).

Change-Id: I7ab06dc9cd7ddff6ff6654d6d611870520abc02b