Add LogeventsPageGenerator, LogpagesPageGenerator
LogpagesPageGenerator ported from compat to be backwards compatible.
LogpagesPageGenerator is marked as deprecated.  number, mode and repeat
are all deprecated: number was changed to total, mode was changed to
logtype, and repeat was removed all-together.

LogeventsPageGenerator is a new generator providing the same
functionality, with updated parameters for core.

Using title() on the results of logitems() returns a Page
(this DOES return a Page but it seems to be a bug. See issue T78125).

Added namespace argument to logevents() which is passed to the request
as "lenamespace".

The command line arguments were also ported from compat using the same
rules. Added TODO for validating log type. This should raise an error.

Unittests in tests/pagegenerators_test.TestLogpagesFactoryGenerator were
added for LogeventsPageGenerator and the command line args. Added
user = True for this test case for testing the use of users in the
LogpagesPageGenerator command line args. These tests have a TODO
indicating the need for a way to check if the logs belong to that user.

Bug: T76555
Change-Id: Ide811094e4d50d5a8375a97bf8ead62eba647814