[IMPROV] Pagegens: Use -logevents instead of -*log
With a887aff93c0aa0859a93cac1e57f4738de59787e the LogeventsPageGenerator
was introduced which uses the -*log notation from compat (e.g. -movelog)
but it's not feasable to support this with the upcoming argparse patch,
because it doesn't allow such dynamic arguments.

This is a replacement so that when the argparse patch lands, that the
users hopefully haven't used the old notation. Instead of semicolons
it's using normal commas which are safer to use on the bash (just the
new notation). It also does only allow positive integers as a total
value (both notations).

Change-Id: If6c24333ea892386203dc9a4affff289ef7f05eb