Re-enable allpages langlinks tests
A set of allpages tests have been commented out due to performing badly,
in 2008.  Of those, filtering for langlinks is now a database error
which has been logged as T78276, and is marked as an allowed_failure
due to it working correctly on very small sites.

The other commented out test filters for no langlinks and now works well.

Change-Id: I62e3adadf24c8dc53e548376d59a2924858be615