Added a list of urls upload support
Test images are now stored in data/images. Use _images_dir as the path variable.
upload_image() is now deprecated, use upload_file(file_url) instead.
Bot's parameter url is deprecated to be string, now it's an iterable.
upload.py does not use urlOK() method anymore. The check is moved to main().
To process_filename() added a check whether there is already a page with the
same title. Docstring for __init__() is extended.

While T85708 is not fixed, try to specify filenames with correct extention.

Bug: T57088
Change-Id: I13d469b3ddcac78b02543efe53868a8672966ec0