Positional args support for GeneratorFactory
Some scripts have positional parameters which are used to create
a page generator.  A common example is page names, which could be
written as -page:"..." but the scripts allow to be simple "...".

In order to bring these parameters into the common infrastructure,
support has been added for positional parameters being detected as
belonging to a standard page generator selected by the script.

Change-Id: I93f9289aa93ea7a263cd8347aa78136065cf4e6b