Allow pagegen filtering using namespace names
The ability to filter using namespace names was removed from the
isbn script in 31c07b2, but still exists in other scripts.
Namespace names are more user-friendly as users interact with
these names regularly in titles, while the namespace numbers are
less known, especially namespaces which are not frequently used.

NamespaceFilterPageGenerator behaviour when site is not provided
is modified from 'no validation of int namespaces' to become
'validation using the namespaces of the default site', which will
at least be equivalent for the builtins.

Equality of Namespace(0) with None is no longer supported.

Exceptions raised during Site methods for namespace validation
are now defined behaviour.

Change-Id: Ib6caa11577546e14a69bbd898860843d69d4efb0