Export UserActionRefuse and deprecate PageNotFound
UserActionRefuse is used by pywikibot and scripts checkimages and
script_wui, but both scripts have been broken since August 2cddfcd
as it is not exported as a top level package class.

PageNotFound was used in compat for HTTP 40x responses, and has
never been implemented in core.  Deprecate it and remove uses of it.

Also, tools.ModuleDeprecationWrapper enhanced to allow custom
deprecation messages.

Change-Id: I55b9912c8d606e1c9f12901ade2d277eb2efabbc