use ValueError when hardcoded config is insecure
The library exception UserActionRefuse is used by the library for
action against a target user that are not possible, such as when
attempting to send an email to a user that has email disabled.

The script_wui constructor is using UserActionRefuse when its hardcoded
config values include a page title which does not end in .js or .css,
which is one way to have a semi-protected wiki page.  The compat version
subster.py does not use this exception.

As this exception is raised only when there is a problem with the
hard-coded config value defined in the same module, and backwards
compatibility would require complicating the library exceptions for the
benefit of only one script which is not in widespread use, and any
working script which monkey-patches script_wui.bot_config needs must
already have a valid config in order to bypass this exception and it
extremely unlikely that any working script will be handling this
exception appropriately, the exception raised is simply changed to
ValueError.

Change-Id: Ic3674c3a82484f7d263bae337db9ed6948c8d228