[FIX] QueryGenerator: Allow missing 'query' entry
The QueryGenerator stopped when the result didn't contain a 'query'
entry. If the result doesn't contain anything yet (but needs to be
continued) it's not returning a 'query' entry when used with the
'generator' API feature.

Instead of stopping the iteration when there is not 'query' entry it's
just stopping the iteration if there is no continuation provided.
Because the exact nature why the condition was there in the first place
(it was added in 54bc6aafa591e31274dc3630e9c6c1039f80839b) is not known
the output was raised from 'debug' to 'log'. Prior to that revision no
condition was found in the repository which looks like that.

Bug: T84860
Change-Id: I1f8c4986d69be18134987c951fd65237300e277e